requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# python-dotenv==1.0.0
# pymongo==4.6.0
# motor==3.3.2
# twilio==8.10.0
# openai==1.3.0
# pydantic==2.5.0
# python-multipart==0.0.6
# httpx==0.25.2

fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
twilio==8.10.0
google-generativeai==0.3.2
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2